+++ build.sh
@@ -1,2 +1,2 @@
 #!/bin/bash
-gcc -o executables/newmodel1 main.c src/shader.c src/camera.c src/newmodel.c src/scene.c src/utils.c src/stb_image.c src/glad.c -L/usr/lib -Iinclude -lglfw -lGL -lcglm -lm -ldl -lassimp
+gcc -o executables/newmodel2 main.c src/shader.c src/camera.c src/model.c src/scene.c src/utils.c src/stb_image.c src/glad.c -L/usr/lib -Iinclude -lglfw -lGL -lcglm -lm -ldl -lassimp
