+++ include/entity.h
@@ -1,6 +1,6 @@
 #ifndef ENTITY_H
 #define ENTITY_H
-#include "newmodel.h"
+#include "model.h"
 #include "shader.h"
 #include <cglm/cglm.h>
