+++ include/model.h
@@ -1,28 +1,51 @@
-#ifndef MODEL_H
-#define MODEL_H
+#ifndef NEWMODEL_H
+#define NEWMODEL_H
 
-#include "mesh.h"
+#include <glad/glad.h>
+#include <cglm/cglm.h>
+#include <stdbool.h>
 #include <assimp/scene.h>
 #include <assimp/postprocess.h>
 #include <assimp/cimport.h>
-#include <string.h>
-#include <stdbool.h>
+#include "shader.h"
 
+typedef enum {
+  TEXTURE_TYPE_DIFFUSE,
+  TEXTURE_TYPE_NORMAL,
+  TEXTURE_TYPE_METALLIC,
+  TEXTURE_TYPE_ROUGHNESS,
+  TEXTURE_TYPE_OCCLUSION,
+  TEXTURE_tYPE_EMISSIVE,
+  TEXTURE_TYPE_UNKNOWN
+} TextureType;
+
+typedef struct {
+    vec3 position;
+    vec3 normal;
+    vec2 tex_coord;
+} Vertex;
+
+typedef struct {
+  GLuint id;
+  TextureType type;
+} Texture;
+
+typedef struct {
+  GLuint VAO, VBO, EBO;
+  unsigned int num_indices;
+  //GLuint texture_id;
+} Mesh;
 
 typedef struct {
   Mesh *meshes;
   unsigned int num_meshes;
+  Shader *shader;
   char *directory;
-  Texture *textures_loaded;
-  unsigned int num_textures_loaded;
 } Model;
 
-Model *model_create(char *path);
-void model_draw(Model *model, Shader *shader);
-bool model_load(Model *model, char *path);
-void model_process_node(Model *model, struct aiNode *node, const struct aiScene *scene, unsigned int *index);
-Mesh *model_process_mesh(Model *model, struct aiMesh *mesh, const struct aiScene *scene);
-TextureArray model_load_material_textures(Model *model, struct aiMaterial *mat, enum aiTextureType type, char *type_name);
-unsigned int TextureFromFile(const char *path, const char *directory);
+bool model_load(Model *model, const char *path);
+void model_process_mesh(Model *model, struct aiMesh *ai_mesh, Mesh *dest_mesh);
+void model_draw(Model *model);
+void model_free(Model *model);
 
 #endif
